% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_emotion.R
\name{classify_emotion}
\alias{classify_emotion}
\title{classifies the emotion (e.g. anger, disgust, fear, joy, sadness, surprise) of a set of texts.}
\usage{
classify_emotion(
  textColumns,
  algorithm = "bayes",
  prior = 1,
  verbose = FALSE,
  lang = "en",
  ...
)
}
\arguments{
\item{textColumns}{A \code{data.frame} of text documents listed one per row.}

\item{algorithm}{A \code{string} indicating whether to use the naive \code{bayes} algorithm or a simple \code{voter} algorithm.}

\item{prior}{A \code{numeric} specifying the prior probability to use for the naive Bayes classifier.}

\item{verbose}{A \code{logical} specifying whether to print detailed output regarding the classification process.}

\item{lang}{Language, "en" for English and "pt" for Brazilian Portuguese.}

\item{\dots}{Additional parameters to be passed into the \code{\link{create_matrix}} function.}
}
\value{
Returns an object of class \code{data.frame} with seven columns and one row for each document.
       \item{anger}{The absolute log likelihood of the document expressing an angry sentiment.}
       \item{disgust}{The absolute log likelihood of the document expressing a disgusted sentiment.}
       \item{fear}{The absolute log likelihood of the document expressing a fearful sentiment.}
       \item{joy}{The absolute log likelihood of the document expressing a joyous sentiment.}
       \item{sadness}{The absolute log likelihood of the document expressing a sad sentiment.}
       \item{surprise}{The absolute log likelihood of the document expressing a surprised sentiment.}
       \item{trust}{The absolute log likelihood of the document expressing a trust sentiment.}
       \item{negative}{The absolute log likelihood of the document expressing a negative sentiment.}
       \item{positive}{The absolute log likelihood of the document expressing a positive sentiment.}
       \item{anticipation}{The absolute log likelihood of the document expressing a anticipation sentiment.}
       \item{best_fit}{The most likely sentiment category (e.g. anger, disgust, fear, joy, sadness, surprise) for the given text.}
}
\description{
\code{classify_emotion} Classifies the emotion (e.g. anger, disgust, fear, joy, sadness, surprise) of a set of texts using a naive Bayes classifier trained on Carlo Strapparava and Alessandro Valitutti's \code{\link{emotions}} lexicon.
}
\examples{
# DEFINE DOCUMENTS
documents <- c("I am very happy, excited, and optimistic.",
               "I am very scared, annoyed, and irritated.")

# CLASSIFY EMOTIONS
classify_emotion(documents,algorithm="bayes",verbose=TRUE, lang = "en")

# pt-BR
documentos <- c("Estou muito feliz, animado e otimista.",
               "Estou muito assustado e irritado.")

# CLASSIFY EMOTIONS
classify_emotion(documentos,algorithm="bayes",verbose=TRUE, lang = "pt")


}
\author{
Timothy P. Jurka <tpjurka@ucdavis.edu> and
        Jodavid Ferreira <jdaf1@de.ufpe.br>
}
