% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_polarity.R
\name{classify_polarity}
\alias{classify_polarity}
\title{classifies the polarity (e.g. positive or negative) of a set of texts.}
\usage{
classify_polarity(
  textColumns,
  algorithm = "bayes",
  pstrong = 0.5,
  pweak = 1,
  prior = 1,
  verbose = FALSE,
  lang = "en",
  ...
)
}
\arguments{
\item{textColumns}{A \code{data.frame} of text documents listed one per row.}

\item{algorithm}{A \code{string} indicating whether to use the naive \code{bayes} algorithm or a simple \code{voter} algorithm.}

\item{pstrong}{A \code{numeric} specifying the probability that a strongly subjective term appears in the given text.}

\item{pweak}{A \code{numeric} specifying the probability that a weakly subjective term appears in the given text.}

\item{prior}{A \code{numeric} specifying the prior probability to use for the naive Bayes classifier.}

\item{verbose}{A \code{logical} specifying whether to print detailed output regarding the classification process.}

\item{lang}{Language, "en" for English and "pt" for Brazilian Portuguese.}

\item{\dots}{Additional parameters to be passed into the \code{\link{create_matrix}} function.}
}
\value{
Returns an object of class \code{data.frame} with four columns and one row for each document.
       \item{pos}{The absolute log likelihood of the document expressing a positive sentiment.}
       \item{neg}{The absolute log likelihood of the document expressing a negative sentiment.}
       \item{pos/neg}{The ratio of absolute log likelihoods between positive and negative sentiment scores. A score of 1 indicates a neutral sentiment, less than 1 indicates a negative sentiment, and greater than 1 indicates a positive sentiment.}
       \item{best_fit}{The most likely sentiment category (e.g. positive, negative, neutral) for the given text.}
}
\description{
\code{classify_polarity} Classifies the polarity (e.g. positive or negative) of a set of texts using a naive Bayes classifier trained on Janyce Wiebe's \code{\link{subjectivity}} lexicon.
}
\examples{
# DEFINE DOCUMENTS
documents <- c("I am very happy, excited, and optimistic.",
               "I am very scared, annoyed, and irritated.")

# CLASSIFY POLARITY


}
\author{
Timothy P. Jurka <tpjurka@ucdavis.edu> and
        Jodavid Ferreira <jdaf1@de.ufpe.br>
}
