% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_accent.R
\name{rm_accent}
\alias{rm_accent}
\title{Function to remove accents from words}
\usage{
rm_accent(str, pattern = "all")
}
\arguments{
\item{str}{vector of strings that will have their accents removed.}

\item{pattern}{string vector with one or more elements indicating which accents should be removed. To indicate which accents should be removed, a vector with the symbols should be passed. Example: pattern = c("´", "^") will strip out the acute and caret accents only. Other accepted words: "all" (remove all accents, which are "´", "`", "^", "~", "¨", "ç")}
}
\value{
vector without accent
}
\description{
\code{rm_accent} Function to remove accents from words
}
\examples{
x <- "São Paulo"

y <- rm_accent(x, pattern = "'")
print(y)

y <- rm_accent(x, pattern = "~")
print(y)

y <- rm_accent(x, pattern = "all")
print(y)

}
